# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog test",
        "description": "Blog test",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get list of articles",
                "description": "Returns list of articles",
                "operationId": "getArticlesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get article information",
                "description": "Returns article data",
                "operationId": "getArticleById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/articles/{id}/comment": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Make comment on an article",
                "description": "",
                "operationId": "storeArticleComment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Article": {
                "title": "Article",
                "description": "Article model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "Title",
                        "description": "Title",
                        "type": "string",
                        "example": "Article title"
                    },
                    "body": {
                        "title": "Body",
                        "description": "Body",
                        "type": "string",
                        "example": "Article body"
                    },
                    "cover_image": {
                        "title": "Cover image",
                        "description": "Cover image",
                        "type": "string",
                        "example": "https://via.placeholder.com/800.png"
                    },
                    "cover_thumbnail": {
                        "title": "Cover thumbnail",
                        "description": "Cover thumbnail",
                        "type": "string",
                        "example": "https://via.placeholder.com/200.png"
                    },
                    "views": {
                        "title": "Views",
                        "description": "Views",
                        "type": "integer",
                        "format": "int64",
                        "example": 0
                    },
                    "published_at": {
                        "title": "Published datetime",
                        "description": "Published datetime",
                        "format": "string",
                        "example": "2022-04-05 14:04:34"
                    },
                    "created_at": {
                        "title": "Created datetime",
                        "description": "Created datetime",
                        "format": "string",
                        "example": "2022-04-05 13:04:34"
                    },
                    "updated_at": {
                        "title": "Updated datetime",
                        "description": "Date time it was last updated",
                        "format": "string",
                        "example": "2022-04-05 16:04:34"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Article"
                }
            }
        }
    },
    "tags": [
        {
            "name": "Projects",
            "description": "API Endpoints of Projects"
        }
    ]
}